replace old(pipeline stage) pippy_mnist example
with compile_stage one
