+++ examples/mnist/pippy_mnist.py
@@ -1,56 +1,66 @@
 # Copyright (c) Meta Platforms, Inc. and affiliates
 import argparse
 import os
-from functools import reduce
 
 import torch
+import torch.distributed as dist
 from torch import nn, optim
 from torch.nn.functional import cross_entropy
 from torch.utils.data import DistributedSampler
 from torchvision import datasets, transforms  # type: ignore
 from tqdm import tqdm  # type: ignore
 
+import pippy
 import pippy.fx
-from pippy import run_pippy
-from pippy.IR import MultiUseParameterConfig, Pipe, PipeSplitWrapper, LossWrapper
-from pippy.PipelineDriver import PipelineDriverFillDrain, PipelineDriver1F1B, PipelineDriverInterleaved1F1B, \
-    PipelineDriverBase
-from pippy.events import EventsContext
+from pippy.IR import PipeSplitWrapper, LossWrapper
 from pippy.microbatch import sum_reducer, TensorChunkSpec
-from pippy.visualizer import events_to_json
 
-PROFILING_ENABLED = True
-CHECK_NUMERIC_EQUIVALENCE = True
-
-schedules = {
-    'FillDrain': PipelineDriverFillDrain,
-    '1F1B': PipelineDriver1F1B,
-    'Interleaved1F1B': PipelineDriverInterleaved1F1B,
-}
-
-LR_VERBOSE = 0
 
 pippy.fx.Tracer.proxy_buffer_attributes = True
 
 USE_TQDM = bool(int(os.getenv('USE_TQDM', '1')))
 
 
-def resolve_pg_per_stage(pp_rank):
-    assert pippy.utils.dp_pg_per_pp_rank
-    return pippy.utils.dp_pg_per_pp_rank[pp_rank + 1]  # exclude master
+# Get process group for ranks in a pipeline
+def get_pp_subgroup(args):
+    my_pp_rank = args.rank // args.dp_group_size
+    my_dp_rank = args.rank % args.dp_group_size
+    for dp_rank in range(0, args.dp_group_size):
+        pp_group_ranks = list(
+            range(dp_rank, args.world_size, args.dp_group_size)
+        )
+        pp_group = dist.new_group(ranks=pp_group_ranks)
+        if dp_rank == my_dp_rank:
+            my_pp_group = pp_group
+    print(f"Rank {args.rank} done getting pp group")
+    return my_pp_group, my_pp_rank
 
 
-def run_master(pp_ranks, args):
+# Get DP process group for ranks with the same stage
+def get_dp_subgroup(args):
+    my_pp_rank = args.rank // args.dp_group_size
+    my_dp_rank = args.rank % args.dp_group_size
+    for pp_rank in range(0, args.pp_group_size):
+        dp_group_ranks = list(
+            range(
+                pp_rank * args.dp_group_size, (pp_rank + 1) * args.dp_group_size
+            )
+        )
+        dp_group = dist.new_group(ranks=dp_group_ranks)
+        if pp_rank == my_pp_rank:
+            my_dp_group = dp_group
+    print(f"Rank {args.rank} done getting dp group")
+    return my_dp_group, my_dp_rank
+
+
+def run_worker(args):
     torch.manual_seed(42)
-    MULTI_USE_PARAM_CONFIG = MultiUseParameterConfig.REPLICATE if args.replicate else MultiUseParameterConfig.TRANSMIT
-    print(f'REPLICATE config: {args.replicate} -> {MULTI_USE_PARAM_CONFIG}')
-    print("Using schedule:", args.schedule)
-    print("Using device:", args.device)
 
-    number_of_workers = 3
-    all_worker_ranks = pp_ranks[1:1 + number_of_workers]  # exclude master
-    chunks = len(all_worker_ranks)
-    batch_size = args.batch_size * chunks
+    # Get DP and PP sub process groups
+    dp_group, dp_rank = get_dp_subgroup(args)
+    pp_group, pp_rank = get_pp_subgroup(args)
+
+    batch_size = args.batch_size * args.chunks
 
     transform = transforms.Compose([
         transforms.ToTensor(),
@@ -60,7 +70,7 @@ def run_master(pp_ranks, args):
     train_data = datasets.MNIST('./data', train=True, download=True, transform=transform)
     valid_data = datasets.MNIST('./data', train=False, transform=transform)
 
-    train_sampler = DistributedSampler(train_data, num_replicas=args.dp_group_size, rank=args.rank, shuffle=False,
+    train_sampler = DistributedSampler(train_data, num_replicas=args.dp_group_size, rank=dp_rank, shuffle=False,
                                        drop_last=False)
 
     train_dataloader = torch.utils.data.DataLoader(train_data, batch_size=batch_size, sampler=train_sampler)
@@ -87,30 +97,33 @@ def run_master(pp_ranks, args):
 
     wrapper = OutputLossWrapper(model, cross_entropy)
 
-    pipe = Pipe.from_tracing(wrapper, MULTI_USE_PARAM_CONFIG, output_loss_value_spec=(False, True))
-    pipe.to(args.device)
+    wrapper.to(args.device)
 
     output_chunk_spec = (TensorChunkSpec(0), sum_reducer)
-    pipe_driver: PipelineDriverBase = schedules[args.schedule](pipe, chunks,
-                                                               world_size=len(all_worker_ranks),
-                                                               all_ranks=all_worker_ranks,
-                                                               output_chunk_spec=output_chunk_spec,
-                                                               _record_mem_dumps=bool(args.record_mem_dumps),
-                                                               checkpoint=bool(args.checkpoint))
 
-    pipe_driver.init_data_parallel(dp_group_size=args.dp_group_size, dp_pg_cb=resolve_pg_per_stage)
+    # sample input
+    x = torch.randint(0, 5, (batch_size, 28, 28), device=args.device)
+    target = torch.randint(0, 9, (batch_size, ), device=args.device)
 
-    optimizer = pipe_driver.instantiate_optimizer(optim.Adam, lr=1e-3, betas=(0.9, 0.999), eps=1e-8)
-    lr_sched = pipe_driver.instantiate_lr_scheduler(optim.lr_scheduler.LinearLR, verbose=LR_VERBOSE)
+    stage = pippy.compile_stage(
+        wrapper,
+        pp_rank,
+        args.pp_group_size,
+        args.chunks,
+        args.device,
+        pp_group,
+        [x, target],
+        output_chunk_spec=output_chunk_spec,
+    )
+
+    optimizer = optim.Adam(stage.submod.parameters(), lr=1e-3, betas=(0.9, 0.999), eps=1e-8)
+    # TODO: add back LR scheduler
+    #lr_sched = pipe_driver.instantiate_lr_scheduler(optim.lr_scheduler.LinearLR, verbose=LR_VERBOSE)
 
     loaders = {
         "train": train_dataloader,
         "valid": valid_dataloader
     }
-
-    this_file_name = os.path.splitext(os.path.basename(__file__))[0]
-    pipe_visualized_filename = f"{this_file_name}_visualized_{args.rank}.json"
-    batches_events_contexts = []
 
     for epoch in range(args.max_epochs):
         print(f"Epoch: {epoch + 1}")
@@ -121,47 +134,38 @@ def run_master(pp_ranks, args):
                 x_batch = x_batch.to(args.device)
                 y_batch = y_batch.to(args.device)
                 if k == "train":
-                    pipe_driver.train()
+                    outp = None
                     optimizer.zero_grad()
-                    outp, _ = pipe_driver(x_batch, y_batch)
-                    preds = outp.argmax(-1)
-                    correct = (preds == y_batch).sum()
-                    all = len(y_batch)
-                    epoch_correct += correct.item()
-                    epoch_all += all
+                    if pp_rank == 0:
+                        stage(x_batch)
+                    elif pp_rank == args.pp_group_size - 1:
+                        outp, _ = stage(y_batch)
+                    else:
+                        stage()
                     optimizer.step()
-                else:
-                    pipe_driver.eval()
-                    with torch.no_grad():
-                        outp, _ = pipe_driver(x_batch, y_batch)
+
+                    if outp is not None:
                         preds = outp.argmax(-1)
                         correct = (preds == y_batch).sum()
                         all = len(y_batch)
                         epoch_correct += correct.item()
                         epoch_all += all
+                else:
+                    # TODO: add evaluation support in PiPPy
+                    pass
 
-                if args.visualize:
-                    batches_events_contexts.append(pipe_driver.retrieve_events())
-            print(f"Loader: {k}. Accuracy: {epoch_correct / epoch_all}")
+            if pp_rank == args.pp_group_size - 1 and epoch_all > 0:
+                print(f"Loader: {k}. Accuracy: {epoch_correct / epoch_all}")
 
-            if k == "train":
-                lr_sched.step()
-                if LR_VERBOSE:
-                    print(f"Pipe {pp_ranks} last_lr: {lr_sched.get_last_lr()}")
-                    print(f"Pipe {pp_ranks} state_dict: {lr_sched.state_dict()}")
+            #if k == "train":
+            #    lr_sched.step()
 
-    if args.visualize:
-        all_events_contexts: EventsContext = reduce(lambda c1, c2: EventsContext().update(c1).update(c2),
-                                                    batches_events_contexts, EventsContext())
-        with open(pipe_visualized_filename, "w") as f:
-            f.write(events_to_json(all_events_contexts))
-        print(f"Saved {pipe_visualized_filename}")
     print('Finished')
 
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
-    parser.add_argument('--world_size', type=int, default=int(os.getenv("WORLD_SIZE", 4)))
+    parser.add_argument('--world_size', type=int, default=int(os.getenv("WORLD_SIZE", 3)))
     parser.add_argument('--rank', type=int, default=int(os.getenv("RANK", -1)))
     parser.add_argument('--master_addr', type=str, default=os.getenv('MASTER_ADDR', 'localhost'))
     parser.add_argument('--master_port', type=str, default=os.getenv('MASTER_PORT', '29500'))
@@ -169,19 +173,33 @@ if __name__ == "__main__":
     parser.add_argument('--max_epochs', type=int, default=10)
     parser.add_argument('--batch_size', type=int, default=10)
 
-    parser.add_argument('-s', '--schedule', type=str, default=list(schedules.keys())[0], choices=schedules.keys())
     parser.add_argument('--replicate', type=int, default=int(os.getenv("REPLICATE", '0')))
     parser.add_argument('--cuda', type=int, default=int(torch.cuda.is_available()))
     parser.add_argument('--visualize', type=int, default=0, choices=[0, 1])
-    parser.add_argument('--record_mem_dumps', type=int, default=0, choices=[0, 1])
     parser.add_argument('--checkpoint', type=int, default=0, choices=[0, 1])
-    parser.add_argument('--exclude_master', type=int, default=0, choices=[0, 1])
+    parser.add_argument(
+        "--chunks",
+        type=int,
+        default=4,
+    )
     args = parser.parse_args()
 
-    args.pp_group_size = 4
-
+    args.pp_group_size = 3
     assert args.world_size % args.pp_group_size == 0
-
     args.dp_group_size = args.world_size // args.pp_group_size
 
-    run_pippy(run_master, args)
+    if args.cuda:
+        dev_id = args.rank % torch.cuda.device_count()
+        args.device = torch.device(f"cuda:{dev_id}")
+    else:
+        args.device = torch.device("cpu")
+
+    # Init process group
+    backend = "nccl" if args.cuda else "gloo"
+    dist.init_process_group(
+        backend=backend,
+        rank=args.rank,
+        world_size=args.world_size,
+    )
+
+    run_worker(args)
